+++ src/pylingdocs/data/project_template/hooks/post_gen_project.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 import cookiecutter
+from configparser import ConfigParser
 
 
 PROJECT_DIR = Path.cwd().resolve()
@@ -20,3 +21,13 @@ if "No license" == "{{ cookiecutter.license }}":
 
 for i, file in enumerate({{cookiecutter.files.split(",")}}):
     create_file(file)
+
+config = ConfigParser()
+config.read(PROJECT_DIR / "pylingdocs.cfg")
+
+path = config.get("paths", "cldf")
+if not path.startswith("/"):
+    path = Path("..") / path
+config.set("paths", "cldf", str(path))
+with open(PROJECT_DIR / "pylingdocs.cfg", "w", encoding="utf-8") as configfile:
+    config.write(configfile)