+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Changed
+
+### Fixed
+
+## [0.0.9] -- 2022-09-19
+
+### Changed
 * more informative `check` function, now with offending files and lines
 * `src` and `psrc` in clld table captions
 * lowercase `paths` and `output` sections in config file
@@ -142,7 +148,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release.
 
-[Unreleased]: https://github.com/fmatter/pylingdocs/compare/0.0.8...HEAD
+[Unreleased]: https://github.com/fmatter/pylingdocs/compare/0.0.9...HEAD
+[0.0.9]: https://github.com/fmatter/pylingdocs/releases/tag/0.0.9
 [0.0.8]: https://github.com/fmatter/pylingdocs/releases/tag/0.0.8
 [0.0.7]: https://github.com/fmatter/pylingdocs/releases/tag/0.0.7
 [0.0.6]: https://github.com/fmatter/pylingdocs/releases/tag/0.0.6